+++ postcss.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  plugins: {
+    // 'postcss-import': {},
+    // 'postcss-cssnext': {},
+    'cssnano': {
+      'autoprefixer': {
+        // 'add': true
+      }
+    }
+  }
+}