+++ webpack.config.lib.es5.js
@@ -1,5 +1,4 @@
 const path = require('path')
-const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin')
 const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
 const merge = require('webpack-merge')
 const common = require('./webpack.config.lib.common.js')
@@ -15,7 +14,6 @@ module.exports = merge(common, {
     library: 'mcv-[name]'
   },
   plugins: [
-    new OptimizeCssAssetsPlugin(),
     new UglifyJSPlugin({
       uglifyOptions: {
         ecma: 5